cloud_functions/requirements.txt
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
oauth2client>=4.0.0
google-cloud-logging